Inline manager logic; drop BasePackageManager; slim index to factory
- Remove BasePackageManager and replicate the small spawn/install/
  uninstall/flatten helpers directly in NpmPackageManager and
  PnpmPackageManager (not enough shared code to warrant a base class).
- Rename PackageManager.ts -> types.ts (now types/interfaces only).
- index.ts exports only the getPackageManager factory; type and class
  imports come straight from their files.
