+++ src/packageManagers/PnpmPackageManager.ts
@@ -1,11 +1,23 @@
 import * as childProcess from 'child_process';
-import type { GetProductionDependenciesOptions, InitOptions, PackageManagerName } from './PackageManager';
-import { BasePackageManager } from './PackageManager';
+import { util } from '../util';
+import type { GetProductionDependenciesOptions, InitOptions, InstallOptions, PackageManager, PackageManagerName, UninstallOptions } from './types';
 
-export class PnpmPackageManager extends BasePackageManager {
+export class PnpmPackageManager implements PackageManager {
     public readonly name: PackageManagerName = 'pnpm';
 
-    protected readonly binaryName = 'pnpm';
+    /**
+     * Spawn pnpm and return a promise that completes when it's finished.
+     * Handles the file extension (`.cmd`) required on windows.
+     * @param args - the list of args to pass to pnpm. Any undefined args will be removed from the list, so feel free to use ternary outside to simplify things
+     */
+    private spawn(args: Array<string | undefined>, options?: childProcess.SpawnOptions) {
+        const filteredArgs = args.filter(arg => arg !== undefined) as string[];
+        return util.spawnAsync(
+            util.isWindowsPlatform() ? 'pnpm.cmd' : 'pnpm',
+            filteredArgs,
+            options
+        );
+    }
 
     public async init(options: InitOptions) {
         //`pnpm init` is non-interactive and does not support a `--force` flag, so we ignore `options.force`
@@ -14,6 +26,24 @@ export class PnpmPackageManager extends BasePackageManager {
         ], {
             cwd: options.cwd,
             stdio: 'inherit'
+        });
+    }
+
+    public async install(packages: string[], options: InstallOptions) {
+        await this.spawn([
+            'i',
+            ...packages
+        ], {
+            cwd: options.cwd
+        });
+    }
+
+    public async uninstall(packages: string[], options: UninstallOptions) {
+        await this.spawn([
+            'uninstall',
+            ...packages
+        ], {
+            cwd: options.cwd
         });
     }
 
@@ -49,6 +79,23 @@ export class PnpmPackageManager extends BasePackageManager {
         const projects = JSON.parse(stdout) as any[];
         const thisProject = (options.packageName ? projects.find(project => project.name === options.packageName) : undefined) ?? projects[0];
         const dependencies = this.flattenDependencyTree(thisProject).filter(x => !!x);
+        return dependencies;
+    }
+
+    /**
+     * Flatten dependencies from `pnpm ls --json` into a list of dependency paths. The host
+     * package is included as the first entry.
+     * @param node a project/dependency node from `pnpm ls --json` (has `path` + nested `dependencies`)
+     * @returns list of dependency paths
+     */
+    private flattenDependencyTree(node: any): string[] {
+        const dependencies: string[] = [];
+        if (node) {
+            dependencies.push(node.path);
+            for (const dep of Object.values(node.dependencies ?? {})) {
+                dependencies.push(...this.flattenDependencyTree(dep));
+            }
+        }
         return dependencies;
     }
 }
